+++ src/app/ride/copy/[...id]/page.tsx
@@ -10,14 +10,14 @@ export const dynamic = "force-dynamic";
 
 export const metadata: Metadata = {
   title: `${env.NEXT_PUBLIC_CLUB_SHORT_NAME} Rides`,
-  description: `${env.NEXT_PUBLIC_CLUB_LONG_NAME} User Profile Page`,
 }
 
 export default async function CopyRidePage({ params }: { params: { id: string } }) {
   const id = flattenQuery(params.id);
   const isAdmin = await canUseAction("ADMIN");
+  const isLeader = await canUseAction("LEADER");
 
-  if (!isAdmin) {
+  if (!isLeader) {
     return (
       <MainContent>
         <h1>Not authorised</h1>